5-HuggingFace/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chromadb==0.3.22
click==8.1.7
clickhouse-connect==0.7.0
comm==0.2.1
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.17.1
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
docutils==0.20.1
duckdb==0.10.0
environs==9.5.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.3
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm-watson-machine-learning==1.0.347
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
jaraco.classes==3.3.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
keyring==24.3.0
kiwisolver==1.4.5
langchain==0.0.345
langchain-core==0.0.13
langsmith==0.0.92
lomond==0.3.3
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
minio==7.2.4
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nh3==0.2.15
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
orjson==3.9.14
packaging==23.2
pandas==1.5.3
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==4.2.0
posthog==3.4.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pycryptodome==3.20.0
pydantic==1.10.14
pydantic_core==2.16.2
pydub==0.25.1
Pygments==2.17.2
pymilvus==2.3.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
readme-renderer==42.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.3
semantic-version==2.10.0
sentence-transformers==2.3.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
toolz==0.12.1
torch==2.2.0
tornado==6.4
towhee==1.1.3
towhee.models==1.1.3
tqdm==4.66.2
traitlets==5.14.1
transformers==4.37.2
triton==2.2.0
twine==5.0.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==11.0.3
wget==3.2
widgetsnbextension==4.0.10
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0
